Fixed tests(hopefully)
